[gcc] Disable long double for IA32. Disable i387 for x86-64.
Rationale: PNaCl can not offer 'long double' since this mode is only supported
by x86 architecture. To make further transition easier we'l disable long double
in default NaCl compiler. Unfortunatelly we can not disable i387 mode for IA32
because SSE is not enough to implement double precision floating point
operations and SSE2 is not implemented in Pentium !!! and Athlon XP (still
widely used in some regions). i387 mode can only ever use 80bit floats
internally - so while we are removing 80bit 'long double' type we are keeping
80bit internal XFmode around.

Also we are disabling SSE math for IA32 - this is not a good default mode for
general code: because SSE only supports single-precision float operations,
double-precision doubles can only be operated with i387 code and switch between
i387 and SSE subsystems is slow on a lot of CPUs. The SSE-base code still can be
used for carefully written codecs and benchmarks, but it's not a default
anymore, it must be enabled explicitly.