[gcc] Make GCC complain if setjmp/longjmp builtins are encountered.
Redefine the builtins in c-torture/.../builtins.exp to make a lot of c-torture
tests pass. This is a dirty way to fix the tests, we will need to enable the
builtins to work with glibc once the latter ready. The purpose of this
manipulation is to reduce the amount of failing tests in the GCC testsuite and
thus make it easier to focus on more interesting failures.

Review URL: http://codereview.chromium.org/2824045