[gcc] Remove few changes from machine-independent part.
This change restores explow.c and reload.c back to their original state.

This change can affect the performance, as it may limit the number of cases when lea instruction is generated.
However, the change is the "right thing" to do. Lea problems, if any, should be solved in i386-specific code.

Review URL: http://codereview.chromium.org/2825071