Fix build regression in commit 18fae96
Commit 18fae96 broke the build with a GCC_NO_EXECUTABLES configure
complaint because it introduced a linking test, which is not allowed.

This redoes that change so it does not overload the native Linux version
and take in its linking test.  The test in question is to determine
whether default linking is behaving as under --enable-new-dtags.
The answer is no with our binutils.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1580
TEST='make build-with-newlib' now works

R=halyavin@google.com,khim@google.com

Review URL: http://codereview.chromium.org/6813007