Restore simple LIB_SPEC using just -lc
The compiler configuration is not the place to encode internal details of
the libc implementation.  When the libc implementation wants the default
link to include multiple constituent libraries, it can do that by providing
a linker script for libc.so or libc.a and let the compiler just use -lc
as it has always done.

This follows a change to the nacl newlib build to provide a libc.a linker
script.  Future changes to the libc implementation internals will not
require any gcc change.

This also removes the truly evil magical treatment of -lnosys as a special
command line option.  It's just a library like any other.

BUG=cleanup
TEST=local builds still link the same using newlib's new libc.a linker script

R=khim@google.com,pasko@google.com,eaeltsin@google.com

Review URL: http://codereview.chromium.org/6995084